+++ packages/account-usage/package.json
@@ -21,7 +21,6 @@
         "@nangohq/shared": "file:../shared",
         "@nangohq/utils": "file:../utils",
         "dd-trace": "5.52.0",
-        "he": "1.2.0",
         "rate-limiter-flexible": "5.0.3",
         "zod": "4.0.5"
     },
