+++ packages/types/lib/plans/http.api.ts
@@ -1,5 +1,5 @@
 import type { BillingCustomer, BillingUsageMetric } from '../billing/types.js';
-import type { MetricUsageSummary } from '../usage/dto.js';
+import type { MetricUsageSummary } from '../usage/index.js';
 import type { ReplaceInObject } from '../utils.js';
 import type { DBPlan } from './db.js';
 import type { Endpoint } from '../api.js';
@@ -55,7 +55,7 @@ export type GetUsage = Endpoint<{
     Path: '/api/v1/plans/usage';
     Querystring: { env: string };
     Success: {
-        data: Record<string, MetricUsageSummary>; //TODO: clean type
+        data: Record<string, MetricUsageSummary>;
     };
 }>;
