--- packages/types/lib/usage/dto.ts
@@ -1,5 +1,0 @@
-export interface MetricUsageSummary {
-    label: string;
-    usage: number;
-    limit: number | null;
-}
